[#9] 하나의 DTO를 요청DTO과 응답DTO 용도로 책임 분산
