+++ ticketing-practice/src/main/java/com/ticketingberry/dto/concert/InConcertDto.java
@@ -1,26 +1,23 @@
-package com.ticketingberry.dto;
+package com.ticketingberry.dto.concert;
 
-import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
-import org.springframework.format.annotation.DateTimeFormat;
-
-import com.ticketingberry.domain.entity.Concert;
+import com.ticketingberry.dto.district.InDistrictDto;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
 
 @Getter
-@Builder
+@SuperBuilder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ConcertDto {
+public class InConcertDto extends ConcertDto {
 	@NotNull(message = "공연이 열리는 장소를 선택해주세요.")
 	private Long placeId;
 	
@@ -28,37 +25,6 @@ public class ConcertDto {
 	private Long artistId;
 	
 	@NotNull(message = "공연에 속하는 구역들을 추가해주세요.")
-	private List<@Valid DistrictDto> districtDtos;	// In
-	
-	@NotNull(message = "제목을 입력해주세요.")
-	@Size(min = 1, max = 150, message = "제목을 1 ~ 50자 사이로 입력해주세요.")
-	private String title;
-	
-	@NotNull(message = "내용을 입력해주세요.")
-	@Size(min = 1, max = 6000, message = "내용을 1 ~ 2000자 사이로 입력해주세요.")
-	private String content;
-	
-	@NotNull(message = "공연 예매가 열리는 시간을 선택해주세요.")
-	@DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
-	private LocalDateTime openedTicketAt;
-	
-	@NotNull(message = "공연이 시작하는 시간을 선택해주세요.")
-	@DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
-	private LocalDateTime performedAt;
-	
-	private List<Long> districtIds;		// Out
-	
-	public static ConcertDto of(Concert concert) {
-		return ConcertDto.builder()
-				.placeId(concert.getPlace().getId())
-				.artistId(concert.getArtist().getId())
-				.title(concert.getTitle())
-				.content(concert.getContent())
-				.openedTicketAt(concert.getOpenedTicketAt())
-				.performedAt(concert.getPerformedAt())
-				.districtIds(concert.getDistricts().stream()
-						.map(district -> district.getId())
-						.collect(Collectors.toList()))
-				.build();
-	}
+	@Builder.Default
+	private List<@Valid InDistrictDto> districtDtos = new ArrayList<>();
 }
