+++ ticketing-practice/src/main/java/com/ticketingberry/dto/reservation/ReservationDto.java
@@ -1,0 +1,14 @@
+package com.ticketingberry.dto.reservation;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@SuperBuilder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReservationDto {
+	protected boolean deposited;
+}
