+++ ticketing-practice/src/main/java/com/ticketingberry/dto/seat/SeatDto.java
@@ -1,15 +1,13 @@
-package com.ticketingberry.dto;
-
-import com.ticketingberry.domain.entity.Seat;
+package com.ticketingberry.dto.seat;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
 
 @Getter
-@Builder
+@SuperBuilder
 @NoArgsConstructor
 @AllArgsConstructor
 public class SeatDto {
@@ -18,14 +16,4 @@ public class SeatDto {
 	
 	@NotNull(message = "좌석 번호를 입력해주세요.")
 	private int seatNum;
-	
-	private Long districtId;	// Out
-	
-	public static SeatDto of(Seat seat) {
-		return SeatDto.builder()
-				.rowNum(seat.getRowNum())
-				.seatNum(seat.getSeatNum())
-				.districtId(seat.getDistrict().getId())
-				.build();
-	}
 }
